user-service/app/models/token_model.py
from sqlmodel import SQLModel, Field
from typing import Optional
from pydantic import BaseModel

class Token(BaseModel):
    access_token: str
    token_type: str
    refresh_token: str


class TokenData(BaseModel):
    username: str | None = None
    email: str | None = None

    