user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Saad-Kamran-2006 <saadkamran6ft@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
sqlmodel = "^0.0.22"
uvicorn = "^0.30.6"
psycopg = "^3.2.1"
httpx = "^0.27.2"
pytest = "^8.3.2"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
types-passlib = "^1.7.7.20240819"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
types-python-jose = "^3.3.4.20240106"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"
psycopg2 = "^2.9.9"
itsdangerous = "^2.2.0"
aiokafka = "^0.11.0"
confluent-kafka = "^2.5.3"
protobuf = "^5.28.0"
types-confluent-kafka = "^1.2.2"
types-beautifulsoup4 = "^4.12.0.20240907"
types-protobuf = "^5.27.0.20240907"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
